feature: add exception - #34

+++ src/main/java/com/example/airplaneletter/service/PostService.java
@@ -1,6 +1,9 @@
 package com.example.airplaneletter.service;
 
 import com.example.airplaneletter.dto.PostDto;
+import com.example.airplaneletter.errorCode.ErrorCode;
+import com.example.airplaneletter.exception.NotFoundException;
+import com.example.airplaneletter.exception.UnauthorizedException;
 import com.example.airplaneletter.model.Post;
 import com.example.airplaneletter.model.User;
 import com.example.airplaneletter.repository.PostRepository;
@@ -31,7 +34,7 @@ public class PostService {
                     .content(post.getContent())
                     .writer(post.getWriter().getNickname())
                     .createdAt(post.getCreatedAt())
-                    .isMyPost(true)
+                    .isMyPost(isPostOwner(user, post))
                     .build();
 
             postList.add(postData);
@@ -52,23 +55,27 @@ public class PostService {
                 .title(postDto.getTitle())
                 .content(postDto.getContent())
                 .nickname(post.getWriter().getNickname())
+                .isMyPost(true)
                 .build();
         return postResponseData;
     }
     public void deletePost(User user, UUID postId) {
         // 포스트 삭제하기.
         Post oldPost = postRepository.findPostById(postId);
+        if(oldPost == null) {
+                throw new NotFoundException(ErrorCode.POST_NOT_FOUND, "해당 post 를 찾을 수 없습니다.");
+        }
 
         if (isPostOwner(user, oldPost)) {
             this.postRepository.delete(oldPost);
         } else {
-            throw new RuntimeException("해당 게시글을 삭제할 수 없습니다.");
+            throw new UnauthorizedException(ErrorCode.FORBIDDEN_USER, "해당 게시글을 삭제할 수 없습니다.");
         }
     }
     public DetailedPostResponseData getPostDetails(User user, UUID postId){
         // 특정 포스트 조회하기.
         Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new IllegalArgumentException("Post not found."));
+                .orElseThrow(() -> new NotFoundException(ErrorCode.POST_NOT_FOUND, "해당 post 를 찾을 수 없습니다."));
         boolean isMyPost = isPostOwner(user, post);
 
         return DetailedPostResponseData.builder()
